backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pydantic
